Struct winch_codegen::FuncEnv

pub struct FuncEnv<'a, 'translation: 'a, 'data: 'translation, P: PtrSize> {
    pub vmoffsets: &'a VMOffsets<P>,
    pub translation: &'translation ModuleTranslation<'data>,
    pub types: &'translation ModuleTypesBuilder,
    pub builtins: &'translation mut BuiltinFunctions,
    /* private fields */
}

The function environment.

Contains all information about the module and runtime that is accessible to to a particular function during code generation.

Fields
vmoffsets: &'a VMOffsets<P>

Offsets to the fields within the VMContext ptr.

translation: &'translation ModuleTranslation<'data>

Metadata about the translation process of a WebAssembly module.

types: &'translation ModuleTypesBuilder

The module’s function types.

builtins: &'translation mut BuiltinFunctions

The built-in functions available to the JIT code.

Implementations

impl<'a, 'translation, 'data, P: PtrSize> FuncEnv<'a, 'translation, 'data, P>

pub fn new( vmoffsets: &'a VMOffsets<P>, translation: &'translation ModuleTranslation<'data>, types: &'translation ModuleTypesBuilder, builtins: &'translation mut BuiltinFunctions, isa: &dyn TargetIsa, ptr_type: WasmValType ) -> Self

Create a new function environment.

pub fn resolve_global(&mut self, index: GlobalIndex) -> GlobalData

Resolves GlobalData of a global at the given index.

pub fn resolve_table_data(&mut self, index: TableIndex) -> TableData

Returns the table information for the given table index.

pub fn resolve_heap(&mut self, index: MemoryIndex) -> HeapData

Resolve a HeapData from a MemoryIndex.

pub fn table_plan(&mut self, index: TableIndex) -> &TablePlan

Get a TablePlan from a TableIndex.

pub fn heap_access_spectre_mitigation(&self) -> bool

Returns true if Spectre mitigations are enabled for heap bounds check.

pub fn name_builtin( &mut self, builtin: BuiltinFunctionIndex ) -> UserExternalNameRef

Creates a name to reference the builtin provided.

pub fn name_wasm(&mut self, index: FuncIndex) -> UserExternalNameRef

Creates a name to reference the wasm function index provided.

pub fn take_name_map( &mut self ) -> PrimaryMap<UserExternalNameRef, UserExternalName>

Extracts the name map that was created while translating this function.