Trait winch_codegen::isa::TargetIsa

pub trait TargetIsa: Send + Sync {
    fn name(&self) -> &'static str;
    fn triple(&self) -> &Triple;
    fn compile_function(
        &self,
        sig: &FuncType,
        body: &FunctionBody<'_>,
        validator: FuncValidator<ValidatorResources>
    ) -> Result<MachBufferFinalized<Final>>;

    fn call_conv(&self) -> CallConv { ... }
    fn endianness(&self) -> Endianness { ... }
}

A trait representing commonalities between the supported instruction set architectures.

Required Methods

fn name(&self) -> &'static str

Get the name of the ISA.

fn triple(&self) -> &Triple

Get the target triple of the ISA.

fn compile_function(
&self,
sig: &FuncType,
body: &FunctionBody<'_>,
validator: FuncValidator<ValidatorResources>
) -> Result<MachBufferFinalized<Final>>

Provided Methods

fn call_conv(&self) -> CallConv

Get the default calling convention of the underlying target triple.

fn endianness(&self) -> Endianness

Get the endianess of the underlying target triple.